[ENG-2961] Fix eslint errors for web extension
