+++ src/app/hooks/queries/ordinals/useInscriptionDetails.ts
@@ -1,16 +1,19 @@
-import { useQuery } from '@tanstack/react-query';
 import useOrdinalsApi from '@hooks/useOrdinalsApi';
+import { useQuery } from '@tanstack/react-query';
 
 const useInscriptionDetails = (id: string) => {
   const OrdinalsApi = useOrdinalsApi();
 
-  return useQuery([`inscription-${id}`], async () => {
-    try {
-      const response = await OrdinalsApi.getInscription(id);
-      return response;
-    } catch (err) {
-      return Promise.reject(err);
-    }
+  return useQuery({
+    queryKey: [`inscription-${id}`],
+    queryFn: async () => {
+      try {
+        const response = await OrdinalsApi.getInscription(id);
+        return response;
+      } catch (err) {
+        return Promise.reject(err);
+      }
+    },
   });
 };
