+++ src/app/screens/forgotPassword/index.tsx
@@ -1,5 +1,6 @@
-import BackHeader from '@components/backHeader';
+import ActionButton from '@components/button';
 import CheckBox from '@components/checkBox';
+import TopRow from '@components/topRow';
 import useWalletReducer from '@hooks/useWalletReducer';
 import { useState } from 'react';
 import { useTranslation } from 'react-i18next';
@@ -16,7 +17,7 @@ const Container = styled.div((props) => ({
 
 const Paragraph = styled.p((props) => ({
   ...props.theme.body_l,
-  color: props.theme.colors.white['200'],
+  color: props.theme.colors.white_200,
   textAlign: 'left',
   marginTop: props.theme.spacing(12),
 }));
@@ -27,42 +28,19 @@ const BottomContainer = styled.div((props) => ({
 
 const ButtonsContainer = styled.div((props) => ({
   display: 'flex',
-  marginTop: props.theme.spacing(16),
   alignItems: 'center',
   justifyContent: 'space-between',
+  marginTop: props.theme.spacing(16),
+  columnGap: props.theme.spacing(8),
 }));
 
-const ResetButton = styled.button((props) => ({
-  display: 'flex',
-  flexDirection: 'row',
-  justifyContent: 'center',
-  alignItems: 'center',
-  borderRadius: props.theme.radius(1),
-  backgroundColor: props.theme.colors.feedback.error,
-  color: props.theme.colors.white['0'],
-  width: '48%',
-  height: 44,
-  '&:disabled': {
-    opacity: 0.6,
-  },
-}));
-
-const CancelButton = styled.button((props) => ({
-  display: 'flex',
-  flexDirection: 'row',
-  justifyContent: 'center',
-  alignItems: 'center',
-  borderRadius: props.theme.radius(1),
-  backgroundColor: props.theme.colors.background.elevation0,
-  border: `1px solid ${props.theme.colors.background.elevation2}`,
-  color: props.theme.colors.white['0'],
-  width: '48%',
-  height: 44,
-}));
+const StyledTopRow = styled(TopRow)({
+  marginLeft: 0,
+});
 
 function ForgotPassword(): JSX.Element {
   const { t } = useTranslation('translation', { keyPrefix: 'FORGOT_PASSWORD_SCREEN' });
-  const [hasBackedUp, setHasBackedUp] = useState<boolean>(false);
+  const [hasBackedUp, setHasBackedUp] = useState(false);
   const navigate = useNavigate();
   const { resetWallet } = useWalletReducer();
 
@@ -81,7 +59,7 @@ function ForgotPassword(): JSX.Element {
 
   return (
     <Container>
-      <BackHeader headerText={t('TITLE')} onPressBack={onBack} />
+      <StyledTopRow title={t('TITLE')} onClick={onBack} />
       <Paragraph>{t('PARAGRAPH1')}</Paragraph>
       <Paragraph>{t('PARAGRAPH2')}</Paragraph>
       <BottomContainer>
@@ -92,10 +70,13 @@ function ForgotPassword(): JSX.Element {
           onCheck={handleToggleBackUp}
         />
         <ButtonsContainer>
-          <CancelButton onClick={onBack}>Cancel</CancelButton>
-          <ResetButton disabled={!hasBackedUp} onClick={handleResetWallet}>
-            Reset
-          </ResetButton>
+          <ActionButton text={t('CANCEL')} onPress={onBack} transparent />
+          <ActionButton
+            text={t('RESET')}
+            disabled={!hasBackedUp}
+            onPress={handleResetWallet}
+            warning
+          />
         </ButtonsContainer>
       </BottomContainer>
     </Container>
