+++ src/app/screens/restoreFunds/restoreBtc/index.tsx
@@ -1,13 +1,10 @@
-import TopRow from '@components/topRow';
-import useWalletSelector from '@hooks/useWalletSelector';
 import IconBitcoin from '@assets/img/dashboard/bitcoin_icon.svg';
-import {
-  BtcUtxoDataResponse,
-  getBtcFiatEquivalent,
-  NetworkType,
-  satsToBtc,
-  UTXO,
-} from '@secretkeylabs/xverse-core';
+import ActionButton from '@components/button';
+import BottomTabBar from '@components/tabBar';
+import TopRow from '@components/topRow';
+import useNonOrdinalUtxos from '@hooks/useNonOrdinalUtxo';
+import useWalletSelector from '@hooks/useWalletSelector';
+import { getBtcFiatEquivalent, NetworkType, satsToBtc, UTXO } from '@secretkeylabs/xverse-core';
 import {
   getBtcFeesForNonOrdinalBtcSend,
   SignedBtcTx,
@@ -16,25 +13,22 @@ import {
 } from '@secretkeylabs/xverse-core/transactions/btc';
 import { useMutation, useQuery } from '@tanstack/react-query';
 import BigNumber from 'bignumber.js';
+import { useEffect } from 'react';
 import { useTranslation } from 'react-i18next';
 import { useNavigate } from 'react-router-dom';
 import styled from 'styled-components';
-import ActionButton from '@components/button';
-import BottomTabBar from '@components/tabBar';
-import { useEffect } from 'react';
-import useNonOrdinalUtxos from '@hooks/useNonOrdinalUtxo';
 
 const RestoreFundTitle = styled.h1((props) => ({
   ...props.theme.body_l,
-  marginBottom: 32,
-  color: props.theme.colors.white[200],
+  marginBottom: props.theme.spacing(16),
+  color: props.theme.colors.white_200,
 }));
 
 const BtcCard = styled.div((props) => ({
   display: 'flex',
   flexDirection: 'row',
   background: props.theme.colors.background.elevation1,
-  borderRadius: 8,
+  borderRadius: props.theme.radius(1),
   padding: '16px 12px',
 }));
 
@@ -60,20 +54,20 @@ const BtcContainer = styled.div({
   flexDirection: 'column',
 });
 
-const Container = styled.div({
+const Container = styled.div((props) => ({
   display: 'flex',
   flex: 1,
   flexDirection: 'column',
-  marginLeft: 16,
-  marginTop: 32,
-  marginRight: 16,
-});
+  marginLeft: props.theme.spacing(8),
+  marginTop: props.theme.spacing(16),
+  marginRight: props.theme.spacing(8),
+}));
 
-const ButtonContainer = styled.div({
-  marginLeft: 16,
-  marginBottom: 32,
-  marginRight: 16,
-});
+const ButtonContainer = styled.div((props) => ({
+  marginLeft: props.theme.spacing(8),
+  marginBottom: props.theme.spacing(16),
+  marginRight: props.theme.spacing(8),
+}));
 
 function RestoreBtc() {
   const { t } = useTranslation('translation', { keyPrefix: 'RESTORE_BTC_SCREEN' });
