+++ src/locales/en.json
@@ -499,7 +499,9 @@
     "TITLE": "Forgot Password",
     "PARAGRAPH1": "Xverse does not keep a copy of your password. If you're unable to access your account, you will need to reset your wallet and input the seedphrase you used when you generated your wallet.",
     "PARAGRAPH2": "This will reset your wallet from this browser. Make sure you have your seed phrase backed up.",
-    "BACKUP_CHECKBOX_LABEL": "I backed up my seedphrase"
+    "BACKUP_CHECKBOX_LABEL": "I backed up my seedphrase",
+    "CANCEL": "Cancel",
+    "RESET": "Reset"
   },
   "RESTORE_WALLET_SCREEN": {
     "ENTER_SEED_HEADER": "Enter your seedphrase to restore your wallet.",
